+++ src/wl_nlp/wl_token_processing.py
@@ -322,7 +322,7 @@ def wl_process_tokens_collocation_extractor(main, text, token_settings):
 
     return text
 
-def wl_process_tokens_colligation(main, text, token_settings):
+def wl_process_tokens_colligation_extractor(main, text, token_settings):
     text = wl_process_tokens(main, text, token_settings)
 
     # Use tags Only