+++ src/CHANGELOG.md
@@ -34,6 +34,7 @@
 - File Area: Show original file paths instead of modified ones
 - File Area: Update File Area
 - Utils: Update CLTK's Norse (Old) stop word list
+- Work Area: Rename Colligation to Colligation Extractor
 - Work Area: Rename Collocation to Collocation Extractor
 - Work Area: Rename N-gram to N-gram Generator
 - Work Area: Rename Overview to Profiler